Still ashamed for my former sloppiness, made Makefile.base even more bulletproof (perhaps overly so :).  Made a rule such that if PLATFORM is set to unknown, it prints out a message indicating that the platform is not recognized and exits.
In order to make this rule not be the first (and therefore default)
rule in every Makefile that includes Makefile.base, I added a rule
"default:" that depends on "all".  Thus, "all" should be the first
target of every Makefile that we make (this seems to be the convention
we both favor anyway, so hopefully this isn't too much of an
imposition).

Also added a Makefile to the make/ directory that prints out all the
PLATFORM/OS/ARCH information.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@28 3a8e244f-b0f2-452b-bcba-4c88e055c3ca