I believe the nightly regressions are now working well enough that it's safe to send the results to chapel_dev without bothering anyone, so I turned that on by default.  This should result in a regressions test status message showing up in the middle of the night every night.
You can see the running tally of tests run, tests passing,
and tests passing with valgrind at:

	http://insidecray.mw.cray.com/~bradc/chapel.html

It's not all that interesting now, but should look good
as we continue to add tests.  John, if you're reading
this far, I haven't yet had a chance to look into the
one valgrind run that's currently failing to see if
it's a real problem or something else needing suppressing.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@245 3a8e244f-b0f2-452b-bcba-4c88e055c3ca