Implemented code generation of the unary operators.  Added a test for bitwise negation.
Added conditions to the findUnOp and findBinOp
functions so that they can't dereference a
null pointer.

Fixed a bug in my generation of the mod function
resulting from fprintf not printing my "%"
symbol since it is a special character.

Inserted a couple of blank lines because I
accidentally committed a minute ago without
including this message.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@371 3a8e244f-b0f2-452b-bcba-4c88e055c3ca