Merged my prototype AST code with the main compiler.  By default, it is not used.  However, it can be selected using the --newast flag.  Currently this causes the new parser and AST to be used and the AST to be printed out, but no code to be generated or compiled.
Added a few tests that I've been using to monitor its evolution
and ensure that I don't go down the wrong path.

Changed the Makefile structure slightly so that the AST code
could continue to live in its own subdirectory, for organizational
purposes.

Changed the name of my Link class to ILink (Inheritance-based
Link) to avoid conflicts with the previous instance.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@426 3a8e244f-b0f2-452b-bcba-4c88e055c3ca