Big checkin.
This adds the new dispatch mechanism and updates all the surrounding
code to use it.  This is a multiple dispatch mechanism that support
destructuring and multiple inheritance.  It reports ambiguties.

Error reporting for typing has been completely overhauled.
Errors are now reported in a manner compatible with gcc (and
hence emacs compilation).  Errors are reported by location in
the code and include a (configurable) call tree depth to help
the user differentiate versions of generic calls.

The testing system has been changed to do sub_clean before a
run so as to cleanup from the last run (necessary since I am
testing the error system and erroroneous compilations don't necessarily
overwrite the previous tests files).


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@458 3a8e244f-b0f2-452b-bcba-4c88e055c3ca