Changed the enumerated binOpType BINOP_EQUALS to BINOP_EQUAL for naming consistency.
Modified code generation of parentheses.  In the past we were
parenthesizing all code generated binary expressions to ensure
precedence correctness.  Now they are only generated when necessary
or to suppress warnings from the C compiler.

Wrote a test generator which generates all combinations of pairs of
binary operators with and without parentheses.  It generates either
C or Chapel code.  Generated and ran C and Chapel versions of the
test in order to compare output.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@776 3a8e244f-b0f2-452b-bcba-4c88e055c3ca