Prior to this checkin, using the --no-codegen flag would cause the AST to be pretty-printed to the console in lieu of code generation.  This was a handy debugging tool early on, but has become increasingly bothersome.  Now, --no-codegen simply means don't do code generation.
Added a pass PrintProgram which will print the program in the
same way that --no-codegen used to.  Set up the
test/trivial/protoast subdirectory to run this pass and only
this pass, since that's what this directory has been testing
up to this point ("can we parse and pretty-print out some
substantial programs").

Modified John's flowanalysis test output, which also used
--no-codegen, so that the pretty-printed output is no
longer intermixed with his analysis pass' output.

Moved the printAST.cpp file from passes to traversals, since
it actually is a traversal (I mistakenly put it in the wrong
place last week).


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@792 3a8e244f-b0f2-452b-bcba-4c88e055c3ca